// src/app/models/saving.ts
export class Saving {
    id: string;
    name: string;
    date: any;
    number: string;
    onlineid: string;
    passwordhint: string;
    memorableplace: string;
    memorablename: string;
    memorabledate: string;
    initialcash: number;
    balance: number;
    interest: number;
    frequency: string;
    annualinterest: number;
    taxable: string;
    notes: string;
    forecast: boolean;
}